refactor(agent): split adapter base runtime internals
